+++ src/run.F90
@@ -82,7 +82,7 @@ contains
 subroutine run()
   type(td_type), pointer :: td
   integer :: iunit, i, ierr
-  real(r8) :: x
+  FLOAT :: x
   logical :: log
   character(len=100) :: filename
